refactor: move deletion process details files and refactor the interface
