+++ AdminApi/src/AdminApi/ClientApp/src/app/components/quotas/identity/identity-details/deletion-processes/deletion-process-details/deletion-process-details.component.html
@@ -46,7 +46,7 @@
                 </mat-panel-description>
             </mat-expansion-panel-header>
             <ng-template matExpansionPanelContent *ngIf="identityDeletionProcess.auditLog">
-                <table mat-table class="responsive" [dataSource]="identityDeletionProcessAuditLogs">
+                <table mat-table class="responsive" [dataSource]="identityDeletionProcessAuditLogs!">
                     <ng-container matColumnDef="id">
                         <th mat-header-cell *matHeaderCellDef>ID</th>
                         <td mat-cell *matCellDef="let AuditLog" data-label="ID:">
